KSENROLL-11219 KSENROLL-11331 support ENR FR qualifier and enable tests
Instead of just supporting milestones (M) and release candidates (RC)

Create a two level heirarchy: major - minor

With FR[0-9]+ being the only major qualifier supported right now.

right now if there is a major qualifier it has to be FR and it can optionally
have one minor qualifier.

When selecting the next version the minor qualifier will be incremented if
it exists.  Otherwise the major qualifier will be used.

for KSENROLL-11331 we enable testing by creating some classes that allow a
local svn repository to be created and then run the tests against it
instead of something in the kuali student repository sandbox.

Each temporary repository created is located in the target build directory
and namespaced using the name provided.  This allows the heirarchy to
support one repo and many different working copies.

this is useful if you do a copy operation because you can then checkout
the destination location to make sure the copy worked as expected.

git-svn-id: https://svn.kuali.org/repos/foundation/trunk@29807 eb505253-de9c-4f3b-bf87-3aad84f29c0f